requirementes.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib

-e .